[#87844102] folding rovi img proxy into the rovi component so that it can be moved as one unit to the private repo
