+++ components/apiproxy.js
@@ -1,51 +1,66 @@
 'use strict';
 
 var proxy = require('http-proxy').createProxy();
+var expressProxy = require('express-http-proxy');
 var querystring = require('querystring');
 var roviSig = require('../node_modules/rovi/lib/sig.js')
 var RateLimiter = require('limiter').RateLimiter;
 // Allow one request every 255ms
 var limiter = new RateLimiter(1, 255);
 
-module.exports = function (pattern, target, host, port) {
-  return function(req, res, next) {
-    limiter.removeTokens(1, function() {
-      var url = req.url;
-      if (url.match(pattern)) {
-        // strip off leading path
-        var newpath = url.replace(pattern, '/');
-        // strip off query string
-        newpath = newpath.replace(/\?.*/, '');
+module.exports = {
+  proxy: function (pattern) {
+    return function(req, res, next) {
+      limiter.removeTokens(1, function() {
+        var url = req.url;
+        if (url.match(pattern)) {
+          // strip off leading path
+          var newpath = url.replace(pattern, '/');
+          // strip off query string
+          newpath = newpath.replace(/\?.*/, '');
 
-        var query = req.query;
-        // rewrite query parameters with ROVI signature
-        query.sig = roviSig(process.env.ROVI_SEARCH_KEY, process.env.ROVI_SEARCH_SECRET);
-        query.apikey = process.env.ROVI_SEARCH_KEY;
-        query.format = 'json'
-        req.url = newpath + '?' + querystring.stringify(query);
+          var query = req.query;
+          // rewrite query parameters with ROVI signature
+          query.sig = roviSig(process.env.ROVI_SEARCH_KEY, process.env.ROVI_SEARCH_SECRET);
+          query.apikey = process.env.ROVI_SEARCH_KEY;
+          query.format = 'json'
+          req.url = newpath + '?' + querystring.stringify(query);
 
-        console.log('proxying', url, req.url);
+          console.log('proxying', url, req.url);
 
-        proxy.on('proxyRes', function (res) {
-          // aggro cache, see https://devcenter.heroku.com/articles/increasing-application-performance-with-http-cache-headers
-          res.headers['Cache-Control'] = 'public, max-age=31536000';
-          res.headers['Expires'] = 'Mon, 25 Jun 2015 21:31:12 GMT';
-        });
+          proxy.on('proxyRes', function (res) {
+            // aggro cache, see https://devcenter.heroku.com/articles/increasing-application-performance-with-http-cache-headers
+            res.headers['Cache-Control'] = 'public, max-age=31536000';
+            res.headers['Expires'] = 'Mon, 25 Jun 2015 21:31:12 GMT';
+          });
 
-        proxy.on('error', function (err) {
-          console.error('proxy error', req.url, err);
-        });
+          proxy.on('error', function (err) {
+            console.error('proxy error', req.url, err);
+          });
 
-        proxy.web(req, res, {
-          target: 'http://api.rovicorp.com',
-          headers: {
-            port: 80,
-            host: 'api.rovicorp.com'
-          }
-        });
-      } else {
-        next();
+          proxy.web(req, res, {
+            target: 'http://api.rovicorp.com',
+            headers: {
+              port: 80,
+              host: 'api.rovicorp.com'
+            }
+          });
+        } else {
+          next();
+        }
+      });
+    }
+  },
+  imgProxy: function() {
+    return expressProxy('cps-static.rovicorp.com', {
+      filter: function(req, res) {
+        return req.method == 'GET';
+      },
+      forwardPath: function(req, res) {
+        res.setHeader('Cache-Control', 'public, max-age=31536000');
+        res.setHeader('Expires', 'Mon, 25 Jun 2015 21:31:12 GMT');
+        return require('url').parse(req.url).path;
       }
-    });
+    })
   }
 };
